# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CorpoPay API",
    "version": "0.1.0",
    "description": "Multi-tenant payment orchestration API."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "InstallmentPlanListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "durationMonths": {
            "type": "number"
          },
          "annualInterestRate": {
            "type": "number"
          },
          "minAmount": {
            "type": "number",
            "nullable": true
          },
          "maxAmount": {
            "type": "number",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "agreementCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "durationMonths",
          "annualInterestRate",
          "minAmount",
          "maxAmount",
          "isActive",
          "agreementCount",
          "createdAt"
        ]
      },
      "InstallmentPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "durationMonths": {
            "type": "number"
          },
          "annualInterestRate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ]
          },
          "minAmount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "nullable": true
              }
            ]
          },
          "maxAmount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "nullable": true
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "name",
          "durationMonths",
          "annualInterestRate",
          "minAmount",
          "maxAmount",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/public/installment-plans/{slug}": {
      "get": {
        "operationId": "publicInstallmentPlans",
        "summary": "List active installment plans for a checkout link",
        "tags": [
          "Public"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "currency": {
                      "type": "string"
                    },
                    "principal": {
                      "anyOf": [
                        {
                          "type": "number"
                        },
                        {
                          "type": "string"
                        }
                      ]
                    },
                    "plans": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "planId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "durationMonths": {
                            "type": "number"
                          },
                          "annualInterestRate": {
                            "type": "number"
                          },
                          "installmentAmount": {
                            "type": "number"
                          },
                          "totalAmount": {
                            "type": "number"
                          },
                          "totalInterest": {
                            "type": "number"
                          },
                          "minAmount": {
                            "type": "number",
                            "nullable": true
                          },
                          "maxAmount": {
                            "type": "number",
                            "nullable": true
                          }
                        },
                        "required": [
                          "planId",
                          "name",
                          "durationMonths",
                          "annualInterestRate",
                          "installmentAmount",
                          "totalAmount",
                          "totalInterest",
                          "minAmount",
                          "maxAmount"
                        ]
                      }
                    }
                  },
                  "required": [
                    "currency",
                    "principal",
                    "plans"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Not an installment link"
          },
          "404": {
            "description": "Payment link not found"
          }
        }
      }
    },
...
    "/installment-plans": {
      "get": {
        "operationId": "listInstallmentPlans",
        "summary": "List installment plans",
        "tags": [
          "Installment Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/InstallmentPlanListItem"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createInstallmentPlan",
        "summary": "Create an installment plan",
        "tags": [
          "Installment Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "durationMonths": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 120
                  },
                  "annualInterestRate": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "minAmount": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "maxAmount": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "name",
                  "durationMonths",
                  "annualInterestRate"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentPlan"
                }
              }
            }
          }
        }
      }
    },
    "/installment-plans/{id}": {
      "patch": {
        "operationId": "updateInstallmentPlan",
        "summary": "Update an installment plan",
        "tags": [
          "Installment Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "durationMonths": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 120
                  },
                  "annualInterestRate": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "minAmount": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "maxAmount": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentPlan"
                }
              }
            }
          },
          "404": {
            "description": "Plan not found"
          }
        }
      },
      "delete": {
        "operationId": "deleteInstallmentPlan",
        "summary": "Delete an installment plan",
        "tags": [
          "Installment Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "deleted"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Plan not found"
          },
          "409": {
            "description": "Plan has active agreements"
          }
        }
      }
    },
...
    "/admin/simulation/bnpl/plans/{tenantId}": {
      "get": {
        "operationId": "simulationBnplPlans",
        "summary": "List a tenant's active installment plans",
        "tags": [
          "Admin Simulation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "tenantId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "durationMonths": {
                            "type": "number"
                          },
                          "annualInterestRate": {
                            "type": "number"
                          },
                          "minAmount": {
                            "type": "number",
                            "nullable": true
                          },
                          "maxAmount": {
                            "type": "number",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "durationMonths",
                          "annualInterestRate",
                          "minAmount",
                          "maxAmount"
                        ]
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}